# perf: avoids vec allocation