--- a/src/machine/asm/mod.rs
+++ b/src/machine/asm/mod.rs
@@ -82,12 +82,9 @@ pub extern "C" fn inited_memory(frame_index: u64, machine: &mut AsmCoreMachine) 
     }
 }
 
-fn check_memory(machine: &mut AsmCoreMachine, page_indices: &[u64]) -> Result<(), Error> {
-    if page_indices.is_empty() {
-        return Ok(());
-    }
-    let frame = page_indices[0] >> MEMORY_FRAME_PAGE_SHIFTS;
-    let frame_end = page_indices.last().unwrap() >> MEMORY_FRAME_PAGE_SHIFTS;
+fn check_memory(machine: &mut AsmCoreMachine, page_indices: &(u64, u64)) -> Result<(), Error> {
+    let frame = page_indices.0 >> MEMORY_FRAME_PAGE_SHIFTS;
+    let frame_end = page_indices.1 >> MEMORY_FRAME_PAGE_SHIFTS;
     for i in frame..=frame_end {
         if machine.frames[i as usize] == 0 {
             inited_memory(i, machine);